node.py
import socket
import threading
import json
from blockchain import Blockchain

class Node:
    def __init__(self, port, bootstrap_node=None):
        self.host = '0.0.0.0'
        self.port = port
        self.bootstrap_node = bootstrap_node
        self.nodes = set()
        self.blockchain = Blockchain()
        self.mempool = []

        if bootstrap_node:
            self.connect_to_node(bootstrap_node)

    def connect_to_node(self, node_info):
        # TODO: Implement a method to connect to a new node
        pass

    def add_node(self, node_info):
        # TODO: Implement a method to add a new node to the list of nodes
        pass

    def remove_node(self, node_info):
        # TODO: Implement a method to remove a node from the list of nodes
        pass

    def broadcast_block(self, block):
        # TODO: Implement a method to broadcast a block to all connected nodes
        pass

    def receive_block(self, block):
        # TODO: Implement a method to handle receiving a block
        pass

    def resolve_forks(self):
        # TODO: Implement a simple fork resolution algorithm
        pass

    def handle_new_transaction(self, transaction):
        # TODO: Implement a method to handle a new transaction
        pass

    def start_server(self):
        # TODO: Implement a method to start the node server
        pass

    def stop_server(self):
        # TODO: Implement a method to stop the node server
        pass

# TODO: Implement additional methods as needed for p2p communication and node functionality
