test_node.py
import pytest
from node import Node

@pytest.fixture
def node():
    return Node(port=5000)

def test_node_initialization(node):
    assert node.port == 5000
    assert node.bootstrap_node is None
    assert isinstance(node.nodes, set)
    assert node.nodes == set()
    assert node.mempool == []

def test_connect_to_node(node):
    # Placeholder test for connect_to_node method
    assert node.connect_to_node('127.0.0.1:5001') is None

def test_add_node(node):
    # Placeholder test for add_node method
    node.add_node('127.0.0.1:5001')
    assert '127.0.0.1:5001' in node.nodes

def test_remove_node(node):
    # Placeholder test for remove_node method
    node.add_node('127.0.0.1:5001')
    node.remove_node('127.0.0.1:5001')
    assert '127.0.0.1:5001' not in node.nodes

def test_broadcast_block(node):
    # Placeholder test for broadcast_block method
    assert node.broadcast_block({'data': 'block'}) is None

def test_receive_block(node):
    # Placeholder test for receive_block method
    assert node.receive_block({'data': 'block'}) is None

def test_resolve_forks(node):
    # Placeholder test for resolve_forks method
    assert node.resolve_forks() is None

def test_handle_new_transaction(node):
    # Placeholder test for handle_new_transaction method
    assert node.handle_new_transaction({'data': 'transaction'}) is None

def test_start_server(node):
    # Placeholder test for start_server method
    assert node.start_server() is None

def test_stop_server(node):
    # Placeholder test for stop_server method
    assert node.stop_server() is None

# Additional tests will be added as the Node class methods are implemented.
